+++ transport/tchannel/handler.go
@@ -123,9 +123,14 @@ func (h handler) handle(ctx context.Context, call inboundCall) {
 		call.Response().Blackhole()
 		return
 	}
+
+	clientTimedOut := ctx.Err() == context.DeadlineExceeded
+
 	if err != nil && !responseWriter.IsApplicationError() {
-		if err := call.Response().SendSystemError(getSystemError(err)); err != nil {
-			h.logger.Error("SendSystemError failed", zap.Error(err))
+		sendSysErr := call.Response().SendSystemError(getSystemError(err))
+		if sendSysErr != nil && !clientTimedOut {
+			// only log errors if client is still waiting for our response
+			h.logger.Error("SendSystemError failed", zap.Error(sendSysErr))
 		}
 		return
 	}
@@ -144,11 +149,11 @@ func (h handler) handle(ctx context.Context, call inboundCall) {
 			responseWriter.AddHeader(ErrorMessageHeaderKey, status.Message())
 		}
 	}
-	if err := responseWriter.Close(); err != nil {
-		if err := call.Response().SendSystemError(getSystemError(err)); err != nil {
-			h.logger.Error("SendSystemError failed", zap.Error(err))
+	if reswErr := responseWriter.Close(); reswErr != nil && !clientTimedOut {
+		if sendSysErr := call.Response().SendSystemError(getSystemError(reswErr)); sendSysErr != nil {
+			h.logger.Error("SendSystemError failed", zap.Error(sendSysErr))
 		}
-		h.logger.Error("responseWriter failed to close", zap.Error(err))
+		h.logger.Error("responseWriter failed to close", zap.Error(reswErr))
 	}
 }